Administracion de categorias

+++ app/Http/Controllers/CategoriaController.php
@@ -9,6 +9,8 @@ use DB;
 
 class CategoriaController extends Controller {
 
+
+
 	/**
 	 * Display a listing of the resource.
 	 *
@@ -17,6 +19,11 @@ class CategoriaController extends Controller {
 	public function index()
 	{
 		return view('admin.categoria.crudcategoria');
+	}
+
+	public function viewSubCategoria()
+	{
+		return view('admin.categoria.crud_sub_categoria');
 	}
 
 	/**
@@ -34,16 +41,40 @@ class CategoriaController extends Controller {
 	 *
 	 * @return Response
 	 */
-	public function store()
+	public function store(Request $request)
 	{
+		$estado = 0;
+		$mensaje = "";
+
 		$categoria = new categoria();
 
 		$categoria->nombre = $request->input('nombre');
-		$categoria->url = $request->input('url');
 		$categoria->categoria_id = $request->input('categoria_id');
 
-		$categoria->save();
+		$lcategoria = DB::table('categorias')
+							->where('categoria_id', '=', $categoria->categoria_id)
+							->where('nombre', '=', $categoria->nombre)->get();
+
+
+		if(count($lcategoria)>0){
+			$estado = 1;
+			$mensaje = "La categoría ya existe.";
+        }else{
+			$estado = 0;
+			$mensaje = "Se guardo correctamente la Categoría.";
+			$categoria->save();
+        }
+
+		$res = $array = [
+	    "estado" => $estado,
+	    "mensaje" => $mensaje
+		];
+
+
+		return $res;
 	}
+
+
 
 	/**
 	 * Display the specified resource.
@@ -59,14 +90,16 @@ class CategoriaController extends Controller {
 	}
 	public function showPrincipal()
 	{
-		$categoria = DB::table('categorias')->where('catagoria_id', '=', 1)->get();
+		$categoria = DB::table('categorias')
+			->where('categoria_id', '=', 1)
+			->where('id', '<>', 1)->get();
 
 		return $categoria;
 	}
 
 	public function showSubCategoria($id)
 	{
-		$categoria = DB::table('categorias')->where('catagoria_id', '=', $id)->get();
+		$categoria = DB::table('categorias')->where('categoria_id', '=', $id)->get();
 
 		return $categoria;
 	}
@@ -89,13 +122,38 @@ class CategoriaController extends Controller {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function update($id)
+	public function update(Request $request, $id)
 	{
-		$categoria = categoria::find($id);
-				$categoria->nombre = $request->input('nombre');
-		$categoria->url = $request->input('url');
 
-		$categoria->save();
+
+		
+		$estado = 0;
+		$mensaje = "";
+		$lcategoria = DB::table('categorias')
+							->where('categoria_id', '=', $request->input('categoria_id'))
+							->where('nombre', '=', $request->input('nombre'))
+							->where('id', '<>',$id)
+							->get();
+
+
+		if(count($lcategoria)>0){
+			$estado = 1;
+			$mensaje = "La categoría ya existe.";
+        }else{
+			$estado = 0;
+			$mensaje = "Se actualizo correctamente la Categoría.";
+
+			$categoria = categoria::find($id);
+			$categoria->nombre = $request->input('nombre');
+			$categoria->save();
+        }
+
+		$res = $array = [
+	    "estado" => $estado,
+	    "mensaje" => $mensaje
+		];
+
+		return $res;
 	}
 
 	/**
@@ -106,8 +164,35 @@ class CategoriaController extends Controller {
 	 */
 	public function destroy($id)
 	{
-		$categoria = categoria::find($id);
-		$categoria->delete();
+		$estado = 0;
+		$mensaje = "";
+		
+
+
+		$lcategoria =DB::select('select * from empresas e join categorias c on e.categoria_id = c.id and c.id = :id', ['id' => $id]);
+		$lsubcategoria =DB::select('select * from categorias c join categorias sc on c.id =sc.categoria_id where c.id = :id', ['id' => $id]);
+
+		if(count($lcategoria)>0){
+			$estado = 1;
+			$mensaje = "La categoría ya es usada en empresas.";
+		}elseif (count($lsubcategoria)>0) {
+			$estado = 1;
+			$mensaje = "La categoría cuenta con sub categorias.";
+        }else{
+
+        	$categoria = categoria::find($id);
+			$estado = 0;
+			$mensaje = "Se elimino correctamente la Categoría.";
+			$categoria->delete();
+        }
+
+		$res = $array = [
+	    "estado" => $estado,
+	    "mensaje" => $mensaje,
+	    "Lista" =>$lcategoria
+		];
+
+		return $res;		
 	}
 
 }
